@@ -11,41 +11,41 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "^11.2.13",
-    "@angular/cdk": "^11.2.12",
-    "@angular/common": "~11.2.13",
-    "@angular/compiler": "~11.2.13",
-    "@angular/core": "~11.2.13",
-    "@angular/forms": "~11.2.13",
-    "@angular/material": "^11.2.12",
-    "@angular/platform-browser": "~11.2.13",
-    "@angular/platform-browser-dynamic": "~11.2.13",
-    "@angular/router": "~11.2.13",
-    "apexcharts": "3.26.2",
-    "echarts": "5.1.1",
+    "@angular/animations": "^14.1.2",
+    "@angular/cdk": "^14.1.1",
+    "@angular/common": "~14.1.2",
+    "@angular/compiler": "~14.1.2",
+    "@angular/core": "~14.1.2",
+    "@angular/forms": "~14.1.2",
+    "@angular/material": "^14.1.1",
+    "@angular/platform-browser": "~14.1.2",
+    "@angular/platform-browser-dynamic": "~14.1.2",
+    "@angular/router": "~14.1.2",
+    "apexcharts": "3.35.4",
+    "echarts": "5.3.3",
     "font-awesome": "^4.7.0",
-    "ng-apexcharts": "1.5.9",
-    "ngx-echarts": "6.0.1",
-    "ngx-toastr": "13.2.1",
-    "ngx-trend": "6.1.1",
-    "rxjs": "7.0.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "ng-apexcharts": "1.7.1",
+    "ngx-echarts": "14.0.0",
+    "ngx-toastr": "15.0.0",
+    "ngx-trend": "8.0.0",
+    "rxjs": "7.5.6",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.12",
-    "@angular/cli": "~11.2.12",
-    "@angular/compiler-cli": "~11.2.13",
-    "@angular/language-service": "~11.2.13",
-    "@angularclass/hmr": "^2.1.3",
-    "@types/echarts": "^4.9.7",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "^12.12.64",
-    "codelyzer": "^6.0.0",
+    "@angular-devkit/build-angular": "~14.1.2",
+    "@angular/cli": "~14.1.2",
+    "@angular/compiler-cli": "~14.1.2",
+    "@angular/language-service": "~14.1.2",
+    "@angularclass/hmr": "^3.0.0",
+    "@types/echarts": "^4.9.16",
+    "@types/jasmine": "~4.0.3",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^18.7.1",
+    "codelyzer": "^6.0.2",
     "resize-observer-polyfill": "^1.5.1",
-    "ts-node": "~8.3.0",
+    "ts-node": "~10.9.1",
     "tslint": "~6.1.0",
-    "typescript": "4.1.5"
+    "typescript": "4.7.4"
   }
 }
